2. 변경 또는 추가사항 -exams_detail로 통합돼 url도 변경 -test_exams_update에서 url도 admin-exam-detail로 변경
