+++ apps/exams/urls/admin.py
@@ -9,16 +9,16 @@ from apps.exams.views.admin.deployments_router import (
 from apps.exams.views.admin.deployments_status import (
     AdminExamDeploymentStatusAPIView,
 )
-from apps.exams.views.admin.exams_delete import AdminExamDeleteAPIView
+from apps.exams.views.admin.exams_detail import AdminExamDetailAPIView
 from apps.exams.views.admin.exams_router import AdminExamRouterAPIView
-from apps.exams.views.admin.exams_update import AdminExamUpdateAPIView
 from apps.exams.views.admin.questions_create import AdminExamQuestionCreateAPIView
 from apps.exams.views.admin.questions_delete import AdminExamQuestionDeleteAPIView
 from apps.exams.views.admin.submissions_list import AdminExamSubmissionListAPIView
 
 urlpatterns = [
     path("exams/<int:exam_id>/questions/", AdminExamQuestionCreateAPIView.as_view(), name="admin-exam-question-create"),
-    path("exams/<int:exam_id>/", AdminExamDeleteAPIView.as_view(), name="admin-exam-delete"),
+    # put + delete
+    path("exams/<int:exam_id>/", AdminExamDetailAPIView.as_view(), name="admin-exam-detail"),
     path("exams", AdminExamRouterAPIView.as_view(), name="admin-exams"),
     path("submissions/", AdminExamSubmissionListAPIView.as_view(), name="admin-exam-submission-list"),
     path(
@@ -37,5 +37,4 @@ urlpatterns = [
         AdminExamDeploymentStatusAPIView.as_view(),
         name="admin-exam-deployment-status",
     ),
-    path("exams/<int:exam_id>/update", AdminExamUpdateAPIView.as_view(), name="admin-exam-update"),
 ]
